moved and renamed dpt abstractions
