+++ xknx/knx/__init__.py
@@ -10,28 +10,29 @@ Module for handling KNX primitves.
 from .address import GroupAddress, GroupAddressType, PhysicalAddress
 from .address_filter import AddressFilter
 from .dpt import DPTArray, DPTBase, DPTBinary, DPTComparator, DPTWeekday
-from .dpt_2byte import (
-    DPT2ByteUnsigned, DPT2Ucount, DPTBrightness, DPTColorTemperature,
-    DPTUElCurrentmA)
+from .dpt_1byte_signed import (
+    DPTPercentV8, DPTSignedRelativeValue, DPTValue1Count)
+from .dpt_1byte_uint import (
+    DPTPercentU8, DPTSceneNumber, DPTTariff, DPTValue1Ucount)
+from .dpt_2byte_float import (
+    DPT2ByteFloat, DPTEnthalpy, DPTHumidity, DPTLux, DPTPartsPerMillion,
+    DPTTemperature, DPTVoltage, DPTWsp)
 from .dpt_2byte_signed import (
     DPT2ByteSigned, DPTDeltaTimeHrs, DPTDeltaTimeMin, DPTDeltaTimeMsec,
     DPTDeltaTimeSec, DPTPercentV16, DPTRotationAngle, DPTValue2Count)
-from .dpt_4byte import DPT4ByteSigned, DPT4ByteUnsigned
+from .dpt_2byte_uint import (
+    DPT2ByteUnsigned, DPT2Ucount, DPTBrightness, DPTColorTemperature,
+    DPTUElCurrentmA)
+from .dpt_4byte_float import (
+    DPT4ByteFloat, DPTElectricCurrent, DPTElectricPotential, DPTEnergy,
+    DPTFrequency, DPTHeatFlowRate, DPTLuminousFlux, DPTPhaseAngleDeg,
+    DPTPhaseAngleRad, DPTPower, DPTPowerFactor, DPTPressure, DPTSpeed)
+from .dpt_4byte_int import DPT4ByteSigned, DPT4ByteUnsigned
 from .dpt_date import DPTDate
 from .dpt_datetime import DPTDateTime
-from .dpt_float import (
-    DPT2ByteFloat, DPT4ByteFloat, DPTElectricCurrent, DPTElectricPotential,
-    DPTEnergy, DPTEnthalpy, DPTFrequency, DPTHeatFlowRate, DPTHumidity,
-    DPTLuminousFlux, DPTLux, DPTPartsPerMillion, DPTPhaseAngleDeg,
-    DPTPhaseAngleRad, DPTPower, DPTPowerFactor, DPTPressure, DPTSpeed,
-    DPTTemperature, DPTVoltage, DPTWsp)
 from .dpt_hvac_contr_mode import DPTHVACContrMode
 from .dpt_hvac_mode import DPTControllerStatus, DPTHVACMode, HVACOperationMode
 from .dpt_scaling import DPTAngle, DPTScaling
-from .dpt_signed_relative_value import (
-    DPTPercentV8, DPTSignedRelativeValue, DPTValue1Count)
 from .dpt_string import DPTString
 from .dpt_time import DPTTime
-from .dpt_value_1_ucount import (
-    DPTPercentU8, DPTSceneNumber, DPTTariff, DPTValue1Ucount)
 from .telegram import Telegram, TelegramDirection, TelegramType
