+++ xknx/knx/dpt_1byte_uint.py
@@ -56,6 +56,16 @@ class DPTPercentU8(DPTValue1Ucount):
     unit = "%"
 
 
+class DPTTariff(DPTValue1Ucount):
+    """
+    Abstraction for KNX 1 Octet tariff information.
+
+    DPT 5.006
+    """
+
+    value_max = 254
+
+
 class DPTSceneNumber(DPTValue1Ucount):
     """
     Abstraction for KNX 1 Octet Scene Number.
@@ -88,13 +98,3 @@ class DPTSceneNumber(DPTValue1Ucount):
             return (knx_value,)
         except ValueError:
             raise ConversionError("Cant serialize %s" % cls.__name__, value=value)
-
-
-class DPTTariff(DPTValue1Ucount):
-    """
-    Abstraction for KNX 1 Octet tariff information.
-
-    DPT 5.006
-    """
-
-    value_max = 254
